requirements.txt
xlwt
django
pymysql
django-environ
django-rest-framework
django-ckeditor
django-crontab
psycopg2-binary
pyyaml
jsonfield
django-db-logger
drf-yasg
django-cleanup
DateTimeRange
pytz
django-cors-headers
django-auto-logout
haversine
cron-descriptor
django-smtp-ssl
requests
pyfcm
pandas
twilio
Pillow
stripe
qrcode
weasyprint==52.2